+++ bika/lims/browser/calcs.py
@@ -223,35 +223,6 @@ class ajaxCalculateAnalysisEntry(BrowserView):
         except ValueError:
             # non-float
             Result['formatted_result'] = Result['result']
-        # calculate Dry Matter result
-        # if parent is not an AR, it's never going to be calculable
-        dm = hasattr(analysis.aq_parent, 'getReportDryMatter') and \
-            analysis.aq_parent.getReportDryMatter() and \
-            analysis.getReportDryMatter()
-        if dm:
-            dry_service = self.context.bika_setup.getDryMatterService()
-            # get the UID of the DryMatter Analysis from our parent AR
-            dry_analysis = [a for a in
-                            analysis.aq_parent.getAnalyses(full_objects=True)
-                            if a.getServiceUID() == dry_service.UID()]
-            if dry_analysis:
-                dry_analysis = dry_analysis[0]
-                dry_uid = dry_analysis.UID()
-                # get the current DryMatter analysis result from the form
-                if dry_uid in self.current_results:
-                    try:
-                        dry_result = float(self.current_results[dry_uid])
-                    except:
-                        dm = False
-                else:
-                    try:
-                        dry_result = float(dry_analysis.getResult())
-                    except:
-                        dm = False
-            else:
-                dm = False
-        Result['dry_result'] = dm and dry_result and \
-            '%.2f' % ((Result['result'] / dry_result) * 100) or ''
 
         self.results.append(Result)